Использован ReportGenerator версии 4.7.1

+++ TestInfrastructure/CodeCoverage/ShowCoverageReport.ps1
@@ -5,11 +5,17 @@ $CodeCoverageFolderPath = "TestInfrastructure\CodeCoverage"
 $AnalyzeFolderPath = $CodeCoverageFolderPath + "\TestResults"
 $ReportFolderPath = $CodeCoverageFolderPath + "\CoverageReport"
 
-$ReportGeneratorVersion = "4.6.1"
+$ReportGeneratorVersion = "4.7.1"
 $ReportGeneratorPath = "$env:USERPROFILE\.nuget\packages\reportgenerator\$ReportGeneratorVersion\tools\netcoreapp3.0\ReportGenerator.dll"
 
-Remove-Item "$AnalyzeFolderPath\*" -Force -Recurse -Confirm:$false
-Remove-Item "$ReportFolderPath\*" -Force -Recurse -Confirm:$false
+if (Test-Path $AnalyzeFolderPath)
+{
+	Remove-Item "$AnalyzeFolderPath\*" -Force -Recurse -Confirm:$false
+}
+if (Test-Path $ReportFolderPath)
+{
+	Remove-Item "$ReportFolderPath\*" -Force -Recurse -Confirm:$false
+}
 
 dotnet test --collect:"XPlat Code Coverage" --results-directory:$AnalyzeFolderPath
